# VM OPSwap optimisation