closes #15114
benchmark vs main
```
goos: darwin
goarch: arm64
pkg: github.com/erigontech/erigon/core/vm/runtime
cpu: Apple M3 Max
                 │   old.txt   │              new.txt               │
                 │   sec/op    │   sec/op     vs base               │
EVM_SWAP1/10k-16   63.24µ ± 2%   59.99µ ± 2%  -5.14% (p=0.000 n=10)

                 │   old.txt    │            new.txt             │
                 │     B/op     │     B/op      vs base          │
EVM_SWAP1/10k-16   4.280Ki ± 4%   4.327Ki ± 4%  ~ (p=0.391 n=10)

                 │  old.txt   │            new.txt             │
                 │ allocs/op  │ allocs/op   vs base            │
EVM_SWAP1/10k-16   49.00 ± 0%   49.00 ± 0%  ~ (p=1.000 n=10) ¹
```